% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRPM-package.R
\docType{package}
\name{miRPM-package}
\alias{miRPM}
\alias{miRPM-package}
\title{miRPM: miRNA RPM Normalization}
\description{
Functions for the analysis of count matrices and RPM normalization in miRNA studies.
}
\details{
This package provides lightweight tools to explore, normalize, and analyze
miRNA-seq count matrices using reads per million (RPM) normalization.
}
\author{
Sergio Pérez Oliveira – sergio30po@gmail.com

Manuel Fernández Sanjurjo – fernandezsanmanuel@uniovi.es

Eduardo Iglesias Gutierrez – iglesiaseduardo@uniovi.es
}
